==> code/dataset/0_0_remove_email_question.ipynb <==
import pandas as pd

# load the survey data from csv file
survey_data = pd.read_csv('../../data/sensitive/raw_questions_12_12_24.csv')

# only keep the column 'email'
survey_data_email = survey_data[['email']]

# randomize the order of the rows for anonymity
survey_data_email = survey_data_email.sample(frac=1).reset_index(drop=True)

# remove empty rows
survey_data_email = survey_data_email.dropna().reset_index(drop=True)

# remove emails that are not ending with '@uos.de' or '@uni-osnabrueck.de'
survey_data_email = survey_data_email[survey_data_email['email'].str.endswith(('@uos.de', '@uni-osnabrueck.de'))]

# save the anonymized email data to a new csv file
survey_data_email.to_csv('../../data/sensitive/questions_12_12_24_emails.csv', index=False)

# drop the column 'email'
survey_data = survey_data.drop(columns=['email'])

# save the data without the email column to a new csv file
survey_data.to_csv('../../data/question_survey/questions_12_12_24_noemail.csv', index=False)
